ai-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
anthropic==0.7.8
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
jinja2==3.1.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pillow==10.1.0
requests==2.31.0